+++ botlink/src/registry.rs
@@ -1,5 +1,4 @@
-use crate::repo::entry_id::{EntryIdRepo, RedisEntryIdRepo};
-use crate::repo::game::{GameRepo, RedisGameRepo};
+use crate::repo::{AttachedBotsRepo, EntryIdRepo, RedisAttachedBotsRepo, RedisEntryIdRepo};
 use crate::stream::xread::{RedisXReader, XReader};
 use crossbeam_channel::{unbounded, Receiver, Sender};
 use micro_model_bot::{ComputeMove, MoveComputed};
@@ -7,7 +6,7 @@ use redis_conn_pool;
 use redis_conn_pool::RedisHostUrl;
 
 pub struct Components {
-    pub game_repo: Box<dyn GameRepo>,
+    pub game_repo: Box<dyn AttachedBotsRepo>,
     pub entry_id_repo: Box<dyn EntryIdRepo>,
     pub xreader: Box<dyn XReader>,
     pub compute_move_in: Sender<ComputeMove>,
@@ -25,7 +24,7 @@ impl Default for Components {
 
         let pool = redis_conn_pool::create(RedisHostUrl::default());
         Components {
-            game_repo: Box::new(RedisGameRepo { pool: pool.clone() }),
+            game_repo: Box::new(RedisAttachedBotsRepo { pool: pool.clone() }),
             entry_id_repo: Box::new(RedisEntryIdRepo {
                 pool: pool.clone(),
                 key_provider: crate::repo::redis_keys::KeyProvider::default(),
