+++ botlink/src/repo/redis_keys.rs
@@ -18,4 +18,7 @@ impl KeyProvider {
     pub fn entry_ids(&self) -> String {
         format!("/{}/botlink/entry_ids", (self.0).0)
     }
+    pub fn attached_bots(&self) -> String {
+        format!("/{}/botlink/attached_bots", (self.0).0)
+    }
 }
